// src/static/src/ts/vacancy/types/participant.ts
export type TYPE_PARTICIPANT = {
  account_number: string;
  avatar_src: string;
  full_name: string;
  id: number;
  id_request: number;
  is_approve: number;
  is_interview: number;
  is_view: number;
  position: string;
  reason_reject: null | string;
  role: number;
  status: number;
};